Added basic evaluation of lisp expressions
Added a single root environment and the ability to lookup
a symbol within the environment. The environment is built
using a hash table.

Added some basic built-in functions including math and
type checks.

Added a unit test for the evaluator which required refactoring
simple test to provide a utility function that evaluates code
and optionally checks the result against an expected value string.