Add lambda functions (#4)
Added lambda functions in the form of:

  (lambda formals body1 body2 ...)

To support the creation of the environment for the lambda function added
the function crisp_bind_env() to evaluator. This takes the list of keys
for the environment (i.e the lambda formals), the list of values to
associate with the keys (i.e. the lambda parameters) and stores them in
a newly created child environment of the passed in environment.

Added test coverage for crisp_bind_env() in a newly created test
called evaluator_test.c. To support testing of invalid argument
cases, added the ability to install an error handler in the interpreter.
The evaluator test installs a handler that will be called when invalid
arguments are passed to crisp_bind_env().

Added a lambda type to the value union, with support as_lambda()
and is_lambda() functions.

Added a special form builtin function called lambda that will construct
a new lambda value when it is applied. This is a special form because it
does not evalaute the operands.