Added variable definition via define keyword
define can be used to bind an expression to a name in the top level environment.
Currently only the following two forms are supported:

    syntax: (define var expr)
    syntax: (define var)

The abbreviated lambda syntax is not supported.

Simple Test unit testing framework needed to be modified to allow the interpreter
instance to persist between calls to TEST_EVAL. This is to allow definitions to be first
defined and then evaluated.