Memory (#6)
Memory logging and leak detection.

Added a memory logging feature that optionally logs all allocations and frees.
This allows for debugging of the memory management logic within the program.
Logging must be enabled via a call to memory_install_logging()

Added hash_table_delete() to remove an entry from the hash table.
A 'deleted' entry is replaced with a tombstone value instead. This allows
the linear probing algorithm to work even if an entry is removed. When
the hash is eventually resized, tombstone entries are removed.