Added Basic Mark and Sweep GC
Added a stop the world, mark and sweep garbage collector, that by default runs after each step of the REPL.

Added a base object into all the objects that can be garbage collected (values and
environments). Each object is registered with the interpreter when it is constructed,
and a linked list of all alive objects is maintained.

When the GC is triggered, it walks through all the values in the root environment and
marks them, along with any values or environments reachable from them. The GC then
steps through the linked list of all objects and frees objects that are not marked.

Strings in the string table are currently not being cleaned up.